refactor: ValidatorModule 위치를 view 하위로 위치 이동
