+++ src/view/ValidatorModule.js
